:file_folder: refactor: alterando estrutura de pastas do app.
